blog/requirements.txt
Django==3.2.18
djangorestframework==3.14.0
djoser==2.1.0
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3

